tests/wpt/web-platform-tests/html/semantics/popups/popup-invoking-attribute.tentative.html

<!DOCTYPE html>
<meta charset="utf-8" />
<title>Popup invoking attribute</title>
<link rel="author" href="mailto:masonf@chromium.org">
<link rel=help href="https://open-ui.org/components/popup.research.explainer">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-actions.js"></script>
<script src="/resources/testdriver-vendor.js"></script>

<button togglepopup=p1>Toggle Popup 1</button>
<div popup=popup id=p1>This is popup #1</div>

<style>
  [popup] {
    border: 5px solid red;
    top: 100px;
    left: 100px;
  }
</style>

<script>
  function clickOn(element) {
    const actions = new test_driver.Actions();
    return actions.pointerMove(0, 0, {origin: element})
      .pointerDown({button: actions.ButtonType.LEFT})
      .pointerUp({button: actions.ButtonType.LEFT})
      .send();
  }

  const popup = document.querySelector('[popup]');
  const button = document.querySelector('button');
  let showCount = 0;
  let hideCount = 0;
  popup.addEventListener('show',() => ++showCount);
  popup.addEventListener('hide',() => ++hideCount);

  async function assertState(expectOpen,expectShow,expectHide) {
    assert_equals(popup.matches(':popup-open'),expectOpen,'Popup open state is incorrect');
    await new Promise(resolve => requestAnimationFrame(resolve));
    assert_equals(showCount,expectShow,'Show count is incorrect');
    assert_equals(hideCount,expectHide,'Hide count is incorrect');
  }

  promise_test(async () => {
    showCount = hideCount = 0;
    await assertState(false,0,0);
    await clickOn(button);
    await assertState(true,1,0);
    popup.hidePopup();
    await assertState(false,1,1);
    button.click();
    await assertState(true,2,1);
    popup.hidePopup();
    await assertState(false,2,2);
  }, "Clicking a togglepopup button opens a closed popup");

  promise_test(async () => {
    showCount = hideCount = 0;
    await assertState(false,0,0);
    await clickOn(button);
    await assertState(true,1,0);
    await clickOn(button);
    await assertState(false,1,1);
  }, "Clicking a togglepopup button closes an open popup");
</script>
